fix tests

+++ ddtrace/internal/ci_visibility/_api_client.py
@@ -76,7 +76,12 @@ _API_RESPONSE_CACHE_DIR = os.path.join(os.getcwd(), ".ddtrace_api_cache")
 
 @atexit.register
 def _clean_api_response_cache_dir():
-    shutil.rmtree(_API_RESPONSE_CACHE_DIR)
+    if os.path.exists(_API_RESPONSE_CACHE_DIR):
+        shutil.rmtree(_API_RESPONSE_CACHE_DIR)
+
+
+def _is_response_cache_enabled():
+    return asbool(os.getenv("_DD_CIVISIBILITY_RESPONSE_CACHE_ENABLED", "true").lower())
 
 
 class TestVisibilitySettingsError(Exception):
@@ -297,7 +302,7 @@ class _TestVisibilityAPIClientBase(abc.ABC):
 
     def _read_from_cache(self, cache_key: str) -> t.Optional[t.Dict]:
         """Read cached response if it exists"""
-        if not cache_key:
+        if not cache_key or not _is_response_cache_enabled():
             return None
 
         cache_file = self._get_cache_file_path(cache_key)
@@ -313,7 +318,7 @@ class _TestVisibilityAPIClientBase(abc.ABC):
 
     def _write_to_cache(self, cache_key: str, data: t.Any) -> None:
         """Write successful response to cache"""
-        if not cache_key:
+        if not cache_key or not _is_response_cache_enabled():
             return
         cache_file = self._get_cache_file_path(cache_key)
         try:
@@ -518,7 +523,7 @@ class _TestVisibilityAPIClientBase(abc.ABC):
         return api_settings
 
     def fetch_skippable_items(
-        self, timeout: t.Optional[float] = None, ignore_test_parameters: bool = False
+        self, timeout: t.Optional[float] = None, ignore_test_parameters: bool = False, read_from_cache: bool = True
     ) -> t.Optional[ITRData]:
         if timeout is None:
             timeout = DEFAULT_ITR_SKIPPABLE_TIMEOUT
@@ -547,7 +552,7 @@ class _TestVisibilityAPIClientBase(abc.ABC):
 
         try:
             skippable_response: _SkippableResponse = self._do_request_with_telemetry(
-                "POST", SKIPPABLE_ENDPOINT, payload, metric_names, timeout
+                "POST", SKIPPABLE_ENDPOINT, payload, metric_names, timeout, read_from_cache=read_from_cache
             )
         except Exception:  # noqa: E722
             return None
@@ -590,7 +595,7 @@ class _TestVisibilityAPIClientBase(abc.ABC):
             skippable_items=items_to_skip,
         )
 
-    def fetch_known_tests(self) -> t.Optional[t.Set[TestId]]:
+    def fetch_known_tests(self, read_from_cache: bool = True) -> t.Optional[t.Set[TestId]]:
         metric_names = APIRequestMetricNames(
             count=EARLY_FLAKE_DETECTION_TELEMETRY.REQUEST.value,
             duration=EARLY_FLAKE_DETECTION_TELEMETRY.REQUEST_MS.value,
@@ -614,7 +619,9 @@ class _TestVisibilityAPIClientBase(abc.ABC):
         }
 
         try:
-            parsed_response = self._do_request_with_telemetry("POST", KNOWN_TESTS_ENDPOINT, payload, metric_names)
+            parsed_response = self._do_request_with_telemetry(
+                "POST", KNOWN_TESTS_ENDPOINT, payload, metric_names, read_from_cache=read_from_cache
+            )
         except Exception:  # noqa: E722
             return None
 
@@ -645,7 +652,7 @@ class _TestVisibilityAPIClientBase(abc.ABC):
 
         return known_test_ids
 
-    def fetch_test_management_tests(self) -> t.Optional[t.Dict[TestId, TestProperties]]:
+    def fetch_test_management_tests(self, read_from_cache: bool = True) -> t.Optional[t.Dict[TestId, TestProperties]]:
         metric_names = APIRequestMetricNames(
             count=TEST_MANAGEMENT_TELEMETRY.REQUEST.value,
             duration=TEST_MANAGEMENT_TELEMETRY.REQUEST_MS.value,
@@ -668,7 +675,7 @@ class _TestVisibilityAPIClientBase(abc.ABC):
 
         try:
             parsed_response = self._do_request_with_telemetry(
-                "POST", TEST_MANAGEMENT_TESTS_ENDPOINT, payload, metric_names
+                "POST", TEST_MANAGEMENT_TESTS_ENDPOINT, payload, metric_names, read_from_cache=read_from_cache
             )
         except Exception:  # noqa: E722
             return None
